+++ src/pyqentangle/__init__.py
@@ -1,7 +1,7 @@
 
 from . import quantumstates
 
-from .utils import OutOfRangeException, UnequalLengthException, InvalidQuantumStateException
+from .core.exceptions import OutOfRangeException, UnequalLengthException, InvalidQuantumStateException
 from .core.schmidt import schmidt_decomposition
 from .continuous import continuous_schmidt_decomposition
 from .metrics import entanglement_entropy, participation_ratio, negativity, concurrence, renyi_entanglement_entropy
