+++ src/pyqentangle/core/schmidt.py
@@ -1,4 +1,6 @@
 
+from typing import Literal
+
 import numpy as np
 import tensornetwork as tn
 
@@ -57,7 +59,10 @@ def schmidt_decomposition_tensornetwork(bipartitepurestate_tensor: np.ndarray) -
     return decomposition
 
 
-def schmidt_decomposition(bipartitepurestate_tensor: np.ndarray, approach: str = 'tensornetwork') -> list:
+def schmidt_decomposition(
+        bipartitepurestate_tensor: np.ndarray,
+        approach: Literal["tensornetwork", "numpy"] = 'tensornetwork'
+) -> list:
     """Calculate the Schmidt decomposition of the given discrete bipartite quantum system.
 
     Given a discrete normalized quantum system, given in terms of 2-D numpy array ``bipartitepurestate_tensor``,
